Разделила функцию на модули
