+++ src/filterType.js
@@ -1,0 +1,7 @@
+'use strict';
+
+module.exports = {
+  'ALL': 'filter-popular',
+  'NEW': 'filter-new',
+  'DISCUSSED': 'filter-discussed'
+};
