+++ src/main.js
@@ -3,3 +3,5 @@
 require('./upload');
 require('./resizer');
 require('./pictures');
+
+
