+++ src/variables.js
@@ -1,0 +1,13 @@
+'use strict';
+
+var filtersBlock = document.querySelector('.filters');
+var picturesContainer = document.querySelector('.pictures');
+var pictures = [];
+var filteredPictures = [];
+
+module.exports = {
+  filtersBlock: filtersBlock,
+  picturesContainer: picturesContainer,
+  pictures: pictures,
+  filteredPictures: filteredPictures
+};
